refactor: move bottom actions to component
